In 9u they use a pitching machine. That means the kids are allowed 5 strikes.
They don't count balls. Kids are encouraged to swing at everything. If we don't count balls, you can't be walked.

If a player has 4 strikes and hits a foul ball on the fifth pitch, they remain at bat. And as long as they hit foul balls after that, they remain at bat until a hit or a strike.